1-09

+++ admin/assets/includes/contents/list.php
@@ -12,7 +12,7 @@ if(isset($_GET["delete"])){
 $contentsList = $contents->list();
 ?>
 
-<!-- <script type="text/javascript">
+<script type="text/javascript">
 function  ConfirmDelete()
 {
     var respuesta = confirm("Estas seguro que deseas Eliminar el contenido?")
@@ -22,7 +22,7 @@ function  ConfirmDelete()
         return false;
     }
 }
-</script> -->
+</script> 
 
 <h2 class="mt-4">Lista de contenidos</h2>
 <table class="table table-response table-hover">
@@ -38,12 +38,9 @@ function  ConfirmDelete()
             <tr>
                 <td> <?= $contenido['title']?></td>
                 <td>
-                    <a href="index.php?opcion=contents&action=update&id=<?= $contenido['id'] ?>">
-                        ACTUALIZAR
-                    </a>
-                    <a href="index.php?opcion=contents&action=list&delete=<?= $contenido['id']?>">
-                        BORRAR
-                    </a>
+                        <a href="index.php?opcion=contents&action=update&id=<?= $contenido['id']?>"><button type='button' class='btn btn-dark'>Actualizar</button></a>
+                        <a href="index.php?opcion=contents&action=list&delete=<?= $contenido['id']?>"><button type='button' class='btn btn-danger' onclick="return ConfirmDelete()">Borrar</button></a>
+
                 </td>
